+++ judge/__init__.py
@@ -1,6 +1,6 @@
 """Template for redbot cog"""
 
-__version__ = "1.1.0"
+__version__ = "1.1.1"
 __author__ = "Ruffiana"
 
 from .main import JudgeCog