+++ judge/CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.0] - 2025-01-15
+## [1.1.1] - 2024-12-22
+
+### Changed
+
+- 5% chance for score will result in all 10s or all 0s.
+
+## [1.1.0] - 2024-12-20
 
 ### Changed